+++ etc/delfin/delfin.conf
@@ -8,4 +8,4 @@ connection = sqlite:////var/lib/delfin/delfin.sqlite
 db_backend = sqlalchemy
 
 [scheduler]
-config_path = '/delfin/scheduler_config.json'
+config_path = '/etc/scheduler_config.json'
